feat(history): migrate C7 audit log history to C8 (#851)
* Add HISTORY_AUDIT_LOG entity type with migration support via AuditLogMigrator and AuditLogTransformer
* Modify C8_KEY column from BIGINT to VARCHAR(255) to support composite keys for audit logs (format: "4095-123456")
* Integrate audit log migration as the last step in HistoryMigrator to ensure all referenced entities exist before migration
* test coverage

https://github.com/camunda/camunda-7-to-8-migration-tooling/issues/517
---------

+++ data-migrator/qa/integration-tests/src/test/java/io/camunda/migration/data/qa/history/entity/HistoryProcessInstanceTest.java
@@ -238,7 +238,7 @@ public class HistoryProcessInstanceTest extends HistoryMigrationAbstractTest {
     historyMigrator.setMode(MigratorMode.MIGRATE);
 
     // Mark parent as skipped
-    dbClient.insert(parentInstance.getId(), null, new Date(), IdKeyMapper.TYPE.HISTORY_PROCESS_INSTANCE, "test skip");
+    dbClient.insert(parentInstance.getId(), (Long) null, new Date(), IdKeyMapper.TYPE.HISTORY_PROCESS_INSTANCE, "test skip");
 
     // when - attempt to migrate (sub should be skipped because parent is skipped)
     historyMigrator.migrate();
